--- src/components/portalMenu.css
@@ -1,27 +1,0 @@
-.colorChangerContainer {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  position: absolute;
-  height: fit-content;
-  padding: 16px;
-  align-items: center;
-  border-radius: 24px;
-  transition: all 0.5s;
-  transform: translate(-50%, -50%);
-  background-color: rgba(15, 15, 15, 0.9);
-  box-shadow: 0px 4px 9px 0px rgba(0, 0, 0, 0.35);
-  max-width: 338px;
-  width: fit-content;
-}
-
-.colorChangerContainer > button {
-  outline: none;
-}
-
-.spacer {
-  height: 42px;
-  width: 1px;
-  background-color: #fff;
-  margin: 2px;
-}
